organize auth helper functions
