+++ utils/Auth/formatErrorCodeToMessage.js
